[Bugfix] Compute triton_mrope row offsets in int64
num_tokens * row_stride is int32 arithmetic in the kernel and wraps past
2^31 elements. The MiniMax M3 ViT reaches it: the encoder budget is
raised to fit one max-size mm item (192k post-merge embeds for a
500-frame video), so a worst-case ViT forward packs ~768k pre-merge
patches, and 768k * 3840 (packed qkv row) > 2^31. Without the cast the
kernel wraps to negative row offsets and the server dies with an
illegal memory access during startup profiling, on NVIDIA and AMD
alike.

Regression test uses M3's geometry (head_dim 80, rotary 78, packed
stride 3840) at 559,248 rows: fails without the fix, passes with it.
Full test_mrope suite: 51 passed, zero perf delta on the microbenchmark.

+++ vllm/model_executor/layers/rotary_embedding/mrope.py
@@ -41,9 +41,10 @@ def _triton_mrope_forward(
     # and supports cos and sin cache with shape (3, num_tokens, rotary_dim // 2)
     # instead of (3, bsz, seq_len, head_dim), also supports interleaved rotary
     pid = tl.program_id(0)
-    # locate start address
-    q_ptr = q_ptr + pid * q_token_stride
-    k_ptr = k_ptr + pid * k_token_stride
+    # locate start address (int64: num_tokens * row_stride exceeds 2^31 for
+    # worst-case ViT forwards, e.g. MiniMax M3 packs ~768k patches/item)
+    q_ptr = q_ptr + pid.to(tl.int64) * q_token_stride
+    k_ptr = k_ptr + pid.to(tl.int64) * k_token_stride
 
     # ####################################################################
     # get the cos(mθ_{i...d/2}) and sin(mθ_{i...d/2}) for token position
